Create new classes to abstract out logic in Bobby class.
Ui class to handle all interactions with users.
Storage class to handle loading and saving of data to the Bobby.txt file.
Parser class to handle user inputs.
TaskList class to maintain the lists of tasks and provide method to add, remove and update tasks status.
Remove the echo function in Bobby as it is not used.