Add more OOP principles
Everything was being done in the Bro.java file, so created more classes to handle different things.

The Ui class that handles all the printing required.
Storage which handles saving tasks and then loading them back in when the bot is run again.
Parser which handles parsing of user inputs.
TaskList which handles adding, deleting, marking, unmarking of tasks