Refactor: Add Ui class to handle all interactions with the user.
Previously, the Garfield class handled all interactions with the user such as printing to screen and
scanning for user input.

This makes the Garfield class bloated as it also contained other logic such as parsing the user input commands and
executing the services for those commands.

Let's abstract all user interactions such as the printing of messages to the screen and scanning of user input
into a new ui class.