Add the save function
Chatbot only remembers the list while it is running, and memory is lost once the program exits.

Saving the tasks to the hard drive allows the task list to be retrieved even after the program exits.

Add save function, which saves the task list to the hard drive whenever the task list is updated. Also, add load function which checks if task list data is saved onto the hard disk and loads it upon startup, else creates new file.