Refactor code to use more oop
The code in the main function of Friday.java is very messy due to a lack
of encapsulation of related code.

The code in Friday.java is very unreadable.

Extracting the code for Ui, Command, Parser, Storage and TaskList into
their respective classes.

Encapsulating the code into those respective classes improves
readability.