Package the application as an executable JAR file
The application is not yet distributed as an executable JAR file,
making it less convenient for users to run the application.

Packaging the application as a JAR file simplifies
 distribution and execution for end-users, providing a single,
easily runnable file.

Package the app into an executable JAR file following the provided
guidelines. Ensure that the JAR file can be run from a command line
with `java -jar "{filename}.jar"`. Make necessary adjustments to
the project setup to support JAR packaging.

Creating an executable JAR file streamlines deployment and
execution processes, making it easier for users to run the
application without additional setup.

The JAR file will not be committed to the Git repository.