Feature: Add TaskList class to wrap the task list
Previously, the task list was implemented as an ArrayList object in the Garfield class, and the Garfield class's logic
directly interfaced with it.

A better way would be to encapsulate the ArrayList in a class and provide methods that would handle the details of adding,
deleting, marking and unmarking tasks.

Let's create a TaskList class to do that.