Organise classes in packages.
Colress, Ui, TaskList, Storage and Parser are in the colress package.
All Task classes are in the colress.task package.
All Command classes are in the colress.command package.
All Exception classes are in the colress.exception package.

Update runtest.bat file in order for the code to compile correctly.