Add function to find a task by date and keyword
Previously there was a function to find task by date only. To make the search more effective, we can add a function to find by keyword as well.

Add FINDDATE and FINDKEY commands to the Command enum.
Implement findTasksByDate and findTasksByKeyword methods in the Parser class.
Update parseFindCommand in Parser to handle both finddate and findkey commands.
Modify Bobby class to integrate handling of FINDDATE and FINDKEY commands and update the run method accordingly.