Add GUI
Command Line App is not appealing to many people who may not use the
app just because the user interface (command line is difficult to use)

The command line also makes it difficult to see history of the chat
with the bot because it does not provide a good text field for the user
to use.

Lets add a GUI using JavaFX to our app to allow better interaction
with the user. Build just the basic app (interaction with the user
only, no welcome message) first.

This will make the app more appealing to potential users and make
interaction much easier.