Complete GUI
The GUI did not have a welcome message, so it seemed very abrupt
when the application started.

This may seem confusing to the reader when they use it for the first
time and be unable to understand what to do next.

Let's add some formatting and a welcome message to the app.

This allows the user to have some form of understanding of what the
app is about to do and have a more intuitive start to the application.