Modify JavaFX version to match runtime version
There is a mismatch between the javafx versions in the .xml files
and the build.gradle file.

This causes error messages to show up and may cause issues in
code behaviour down the road.

Lets modify both javafx versions to be '17.0.7'.

This will ensure consistency in javafx versions being used.