Add Data folder to Resources
List Storage file is left as a standalone file inside of the
resources folder.

This is messy for code organisation and readability and reduces
any improvement potential by other developers in future.

Lets create a folder 'data' to keep the storage list file for better
organisation.

This will make the code more readable and intuitive for future
developers.