Add Varargs to find functionality
Find is restricted to only one search phrase.

This makes it inflexible to search for things that may be related
but not with the same phrases. For example, read book and write
report are both within similar domains of study but do not share
similar phrases. As such, there needs to be a way to quick search
them.

Lets add Varargs to the find function, and allow different search
terms to be separated with a comma.

This allows similar types of things to be searched together.