Add dynamic background colour for DialogBox
The background colours for the DialogBox are pretty uninteresting as
they are all green with black text.

The user is unable to immediately tell if the command they sent was
indeed recorded correctly or not as there is no visual feedback.

Lets add dynamic colouring to the dialogbox so that there is visual
feedback on the type of command that was just processed by the user.

This will give the user a better user experience and ability to
instantly recognise if they had made an error in passing a command
previously just by the colour of the dialogbox.