Modify JavaDocs
Some methods' Javadocs were not updated after the GUI was added.

These methods changed from being void (due to the use of System.out.
println) but changed to returning a string which then is printed on
the GUI.

This makes the Javadoc outdated and difficult for future developers
to refer to.

Lets update the Javadocs to make the code more readable and
understandable.

This would then allow future developers to understand the use of the
modified methods with no confusion.