feat: Add GUI for ZBot application
This commit introduces a GUI for the existing ZBot program.

This is the first iteration in GUI  development and it does not possess all the features available in the text-based UI in previous commits.

The user is currently able to `list` and add tasks with `todo`, `deadline`, `event`, `mark`, `unmark`, `delete`, `find`. The `bye` command to exit program feature has not been transferred. User should close the GUI window to terminate the program instead.