Change format for deadlines and event start and end to LocalDateTime
The previous type for storing deadlines and event timings was String, which
does not allow for more meaningful storage.

Changing the type to LocalDateTime allows the chatbot to understand that
it is a date and time instead of just a String.

Let's edit the DeadlineTask and EventTask variables to reflect this, as well
as change the format of parseTask() to support reading such data from files.