Update data file path to absolute path
Currently, the data file path is relative to the project root directory.
This causes the file to be created in the root directory of the project.
This has caused problem in the JAR file as the file destination mightnot be found.

Let's update the data file path to an absolute path to prevent this issue.
The storage test cases also have been tweaked to use absolute path for testing.