Derrick class: add assertions to document important assumptions
There are many assumptions made during the initialization of the
Derrick class.

The assumptions if left unchecked could cause problems in the
future

Adding assertions allow us to ensure that these assumptions are
documented well. For instance, adding assertions to ensure that
both the tasks and storage are initialized correctly.

Let's add assertions to check whether the tasks and storage fields
are initialized correctly