Debug on exit handling for JavaFX
The application does not correctly close when the "bye" command is
issued in the JavaFX GUI.

The improper shutdown handling can leave the JavaFX runtime hanging,
leading to potential resource management issue and a poor user
experience.

Use Platfrom.exit() in the ExitCommand class to properly terminate the
JavaFX application.

Platform.exit() allows for a graceful shutdown of the JavaFX platform.

This changes provide a reliable and user-friendly way to exit the
applicaiton, avoiding abrupt terminations.