Add assertions to methods in command classes
Subclasses of Command do not have any checks and validations
to input arguments.
For example, the method execute expects a TaskList object
to call list operation methods on.

Null arguments could accidentally be passed, and may result
in NullPointerExceptions if we try to call such methods on the object.

By adding assertions, we check that arguments passed are not null.

This allows us to catch if null arguments are accidentally passed.

Let's add assert statements to check if arguments are not null.