Add assertions parse method in Parser
Parser class does not have any checks and validations
to input argument.
In this case, the method expects a string command to parse.

Null argument could accidentally be passed, and may result
in NullPointerExceptions if we try to call String methods on it.

By adding assertions, we check that arguments passed are not null.

This allows us to catch if null arguments are accidentally passed.

Let's add assert statements to check if arguments are not null.