Add assertions to contructor of FileStorage
The constructor does not have any checks and validations
to input argument.
In this case, the constructor expects a path to create the directory.

Null argument could accidentally be passed, and may result
in NullPointerExceptions if we try to create directories or
folders.

By adding assertions, we check that arguments passed are not null.

This allows us to catch if null arguments are accidentally passed.

Let's add assert statements to check if arguments are not null.