Add assertions to constructor of TaskList
The constructor does not have any checks and validations
to input argument.
In this case, the constructor expects a FileStorage object.

Null argument could accidentally be passed, and may result
in NullPointerExceptions if we try to call operations on the object.

By adding assertions, we check that arguments passed are not null.

This allows us to catch if null arguments are accidentally passed.

Let's add assert statements to check if arguments are not null.