Add individual methods to generate corresponding Command object
Command string is parsed in a long chain of if-else in
parseCommand method.

This does not follow SLAP, as it makes 1 method responsible
for checking, parsing, and creating the relevant command object
for all commands. This also makes the metho body very long.

Let's isolate parsing of command string and creation of
Command object for each command to its own method.

This adheres more closely to SLAP.