fix: Fix application unable to close after "bye" command
The application remains open after using the 'bye' command due to the
lack of proper shutdown logic in the GUI.

Users expect the application to close when they use the "bye" command,
the lack of this functionality could causes confusion and a poor user
experience.

Let's,
* add a method in Dude to check if it is running
* add method in MainWindow to close the application after a short delay if
Dude is no longer running

The short delay ensures that application display the exit message before
the it exits.