Add GUI interface
Due to the requirement of this course and make this chatbot easy to
use, added GUI interface.

Add new gui package, collects all the class relevant to javafx gui
interface. Add class Gui, which manage the whole gui, which is
also the stage of javafx. Add class MainWindow, which manage the user
interaction and the response of the chatbot. Add class DialogBox,
easier to manage the dialog box in the user interface. Add resources
directory, manage all the images and the fxml files, make the ui design
become easier.

Collect the classes Gui, MainWindow, DialogBox into one package gui
help to organized the whole project well.