Refactor code for Parser and UpdateTaskCommand classes
There were some areas of code that could be simplified to enhance
readability.

Changes were made to the Parser and UpdateTaskCommand classes to
improve code comprehension for future debugging

The 'createParsedTask' method was extracted from the 'parseFileData'
method to enhance the readability of 'parseFileData'. The code in
the 'execute' method of UpdateTaskCommand was reorganised.