Reformat code
The code is not completely adherent to the code style and
conventions inside of the CS2103 requirements.

This makes the code seem like it is not consistently done by one
person.

Lets modify some parts of the code not adherent to the codestyle.

This will make the code more consistent and readable.