Add assertions
The code assumes that the file name provided to Storage is not empty
and that the Storage object provided to TaskList is not null.

This can lead to exceptions being thrown, which are hard to debug.

Add assertions to verify that such assumptions are true and to provide
informative error messages if such assumptions are false.