Replace some loops with streams
Streams are clearer and show the step
by step process of an action better

Let's
* trim all the trailing spaces at once with
a stream
* add to the option map with a stream

This is done by utilising map functions
to trim and collect the options into an array.
Then, we skip the first option and
utilise the map function to generate
the option map.

In addition, as the method parse is long,
we moved out task parsing logic
out into a private method.