Add assertion at the end of FileStorage creation.
Currently, there are no assertions in the code of
Rizzler, which may lead to assumptions in the code
that turn out to be wrong.

Adding assertions to the code eliminates such
assumptions and improves code quality.

As a step towards including assertions into the
code, let's introduce assertions into the
FileStorage class to assert save file creation.