Extract methods in AddCommand for readability
The codebase has some methods that are considerably lengthy,
which inhibits readability and by extension code quality.

Extracting portions of those methods out into new methods
reduces the length of each method, increasing the readability
of the codebase.

Let's increase the readability of the codebase by extracting
the input checking of AddCommand during command execution into separate
methods to reduce the length of the execute() function.