ModifyCommand class: Add assertion
The if-elseif-else block works as expected only if the assumption holds
true that the code in the else block is executed only if
this.verb == CommandVerb.DELETE.

However, there may be bugs elsewhere in the programme which results
in the ModifyCommand object to be instantiated such that this.verb
accepts a CommandVerb Enum other than MARK, UNMARK or DELETE, rendering
the assumption to be invalid.

Let's add an assert statement to inform programmers when the assumption
fails to hold.

This is to ensure that bugs are caught early on and prevent an
accumulation of bugs as the programme continues to be developed.