Refactor files
The codes are messy and the code quality was not there for some files.

It needs to be more readable.

Let's refactor the modified files to make it more readable.

By being more readable, it will be easier to read and easier for new
programmers to work on this project.