Command class: Remove isExit() function
The code that drives the application in Main.java does not require to
check whether commands parsed is a kind that require exit of
application. This has simplified the codebase.

To increase readability of code, isExit() can be deleted.

Let's remove isExit() in all Command class and subclasses to improve
readability.