Add Assertion to TaskList.java
There is a method deleteTask() which holds an assumption that has
has not been explicitly shown in the code. Delete task will always
only be run once there are tasks that have been added to the
taskList in the first place and have already been checked for
validity in the addTask or loadTask method call.

By not showing this assumption to the future developers, they may
not be aware of this and try to change the method in an unintended
way, which may break the delete function.

Let's create an assertion at the else statement inside the
deleteTask method that ensures the code does not ever reach that
location.

This will help new developers understand the code better AND it
becomes more apparent when this assumption is no longer true,
prompting the developer to modify other methods like the addTask
and loadTask method such that the assumption becomes true again.