Add assertions to Processor.java
There is a fair bit of assumptions made inside the methods of
Processor.java, specifically in the markUnmarkOrDelete and
createTask methods.

In both methods, if-else statements were used to call on the last
conditional in the method but if the handleCommand method was to be
modified, there is no guarantee that this assumption will hold.

Let's add these assumptions to ensure that the developer knows about
the said assumptions and modifies other methods in the project
around these assumptions.

This will make the project less confusing to work on in the future.